Add simple base class for Salesforce KW Libraries
